# Conflicts:
#	angel-ps/mllib/src/main/java/com/tencent/angel/ml/classification/lr/LRLearner.scala
#	angel-ps/mllib/src/main/java/com/tencent/angel/ml/optimizer/sgd/GradientDescent.scala

+++ angel-ps/core/src/main/java/com/tencent/angel/ml/math/vector/CompDoubleVector.java
@@ -740,7 +740,7 @@ abstract class CompDoubleVector extends TIntDoubleVector {
   }
 
   @Override public double sparsity() {
-    return nonZeroNumber() / getDimension();
+    return (double)nonZeroNumber() / (double)getDimension();
   }
 
   @Override public int size() {
